move the components into one file and export connected as default after PR review suggestion
